Former-commit-id: df612396fb55b665a4b837035028652664867aa5 [formerly dce0dc369255772d8d6bfd14aab76a11b4932790] [formerly 097617fa26d848bfd8faf46486002d72ba5b641f] [formerly df612396fb55b665a4b837035028652664867aa5 [formerly dce0dc369255772d8d6bfd14aab76a11b4932790] [formerly 097617fa26d848bfd8faf46486002d72ba5b641f] [formerly 462652e31eaa71a28a54211fff68c397ea4d9e73 [formerly 097617fa26d848bfd8faf46486002d72ba5b641f [formerly eef8340cb8e904e4469e9c99271a34268ca2ce23]]]]
Former-commit-id: 462652e31eaa71a28a54211fff68c397ea4d9e73
Former-commit-id: a6bb8b38dd57977534e76989bb580f6d36f1ecf1 [formerly 026674d0b2431fd636c1eee422b0eaed908af4cf] [formerly 19cf9a00ac458922c15640956ecdf8f46546e349 [formerly 299690a202796c064fb0cad5ab65095a817db326]]
Former-commit-id: 1b7b8110c512177341c86e9272a2b79f0ae52bbb [formerly bd9b2e493ea54d0dceab558f2ca648a2e9e4c5ad]
Former-commit-id: 6dbe80b1aff1d7fe1cdd45cf9ebd3feeeab08234

+++ rpms/awips2.core/Installer.ldm/component.spec
@@ -1,5 +1,7 @@
-%define _ldm_version 6.11.2
-%define _ldm_src_tar awips2-ldm.%{_ldm_version}.tar.gz
+%define _ldm_version 6.11.5
+%define _ldm_src_tar ldm-%{_ldm_version}.tar.gz
+# ldm-%{_ldm_version}.tar.gz is tarred up ldm-%{_ldm_version}/src dir after
+# ISG makes retrans changes
 #
 # AWIPS II LDM Spec File
 #
@@ -18,10 +20,8 @@ Vendor: Raytheon
 Packager: Bryan Kowal
 
 AutoReq: no
-Requires: qpid-cpp-client = 0.7.946106-28.el5.centos.1 
-Requires: qpid-cpp-client-devel = 0.7.946106-28.el5.centos.1
+Requires: awips2-notification
 Requires: zlib-devel
-Requires: /usr/lib/libz.a
 provides: awips2-ldm
 provides: awips2-base-component
 
@@ -51,6 +51,21 @@ if [ $? -ne 0 ]; then
 fi
 
 /bin/mkdir -p %{_build_root}/etc/profile.d
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+
+/bin/mkdir -p %{_build_root}/etc/ld.so.conf.d
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+
+/bin/mkdir -p %{_build_root}/etc/logrotate.d
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+
+/bin/mkdir -p %{_build_root}/etc/init.d
 if [ $? -ne 0 ]; then
    exit 1
 fi
@@ -118,6 +133,21 @@ if [ $? -ne 0 ]; then
    exit 1
 fi
 
+/bin/cp ld.so.conf.d/* %{_build_root}/etc/ld.so.conf.d
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+
+/bin/cp logrotate.d/* %{_build_root}/etc/logrotate.d
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+
+/bin/cp init.d/* %{_build_root}/etc/init.d
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+
 %pre
 if [ -d /tmp/ldm ]; then
    rm -rf /tmp/ldm
@@ -126,7 +156,7 @@ mkdir -p /tmp/ldm
 for dir in etc .ssh;
 do
    if [ -d /usr/local/ldm/${dir} ]; then
-      scp -qrp /usr/local/${dir} /tmp/ldm
+      scp -qrp /usr/local/ldm/${dir} /tmp/ldm
    fi
 done
 
@@ -139,12 +169,12 @@ _myHost=`echo ${_myHost} | cut -f1 -d'-'`
 pushd . > /dev/null 2>&1
 cd ${_ldm_dir}/SOURCES
 # unpack the ldm source
-/bin/tar -xf awips2-ldm.%{_ldm_version}.tar.gz \
+/bin/tar -xf %{_ldm_src_tar} \
    -C ${_ldm_dir}
 if [ $? -ne 0 ]; then
    exit 1
 fi
-rm -f awips2-ldm.%{_ldm_version}.tar.gz
+rm -f %{_ldm_src_tar}
 if [ $? -ne 0 ]; then
    exit 1
 fi
@@ -401,3 +431,6 @@ rm -rf ${RPM_BUILD_ROOT}
 /usr/local/ldm/SOURCES/*
 
 %attr(755,root,root) /etc/profile.d/awipsLDM.csh
+%attr(755,root,root) /etc/ld.so.conf.d/awips2-ldm-i386.conf
+%attr(755,root,root) /etc/ld.so.conf.d/ldm.log
+%attr(755,root,root) /etc/init.d/ldmcp
